Remove auto-generated tests/list_atomic.c and test_atomic_include.h from the repo.
The files are still present in the distributive tar-ball as it is hard
to generate them on Windows (without Cygwin/MinGW).  Note that these
files are used only for testing (unlike generalize-small.h).
The later remains even in the repo as it makes possible to use
libatomic_ops without running any script.

* .gitignore: add tests/list_atomic.c and test_atomic_include.h.
* src/Makefile.msft: Add comment about test_atomic_include.h; wrap
long lines.
* src/Makefile.msft (all): Build "libatomic_ops_gpl.lib" instead of
"check".
* tests/list_atomic.c, tests/test_atomic_include.h: Remove.