Fix 'duplicate symbol' error for test_malloc/stack with static libs (OS X)
If configured with --enable-static (the default behavior), libtool
passes libatomic_ops.a to gcc twice (with a relative path and with
an absolute one) when linking test_malloc and test_stack because
the latter ones use also depend on libatomic_ops_gpl.a which, in turn,
depends on libatomic_ops.a.  Double specification of  libatomic_ops.a
confuses ld tool of MacOS 10.14.

The workaround is to omit libatomic_ops.la in *_LDADD specification
for test_malloc and test_stack at least when the shared libraries are
not requested.

* configure.ac (ENABLE_SHARED): New AM_CONDITIONAL.
* tests/Makefile.am (test_stack_LDADD, test_malloc_LDADD): Do not add
libatomic_ops.la unless ENABLE_SHARED; add comment.