Fix get_chunk for case of mmap area is located before AO_initial_heap (a cherry-pick of commit 14da3ff7d3 from 'release-7_4')
System mmap() might return a memory-mapped region which is located
on addresses smaller than program area for static data.  Thus, value
of my_chunk_ptr in get_chunk() might be expected to be smaller than
the address of AO_initial_heap[0] and this case should be treated
in the same way as if my_chunk_ptr points above
&AO_initial_heap[AO_INITIAL_HEAP_SIZE-CHUNK_SIZE].

* src/atomic_ops_malloc.c (get_chunk): If my_chunk_ptr is less than
(or equal to) AO_initial_heap, then also call get_mmaped.